backend/requirements.txt
Flask
Flask-CORS  # <-- NEW: This is essential for a separate frontend
tensorflow
numpy
Pillow
gunicorn